requirements.txt

streamlit==1.28.1
streamlit-autorefresh==0.0.1
streamlit-components-template==0.0.1
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
spacy==3.7.2
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2
datasets==2.14.5
accelerate==0.21.0
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
datapackage==1.15.2
SPARQLWrapper==2.0.0
wikidata==0.7.0
dbpedia-spotlight==0.4.0
tqdm==4.66.1
zipfile38==0.0.3
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.24.1
